requirements.txt
Flask==1.1.1
requests==2.22.0
pycryptodome==3.9.6
cryptography==3.3.2
responses==0.10.15
pylint==2.13.3
pytest==7.1.1
pytest-cov==3.0.0
wheel==0.35.1
